ngtcp2: share common functionality
Share common functions/structs between ngtcp2 HTTP/3 and the
proxy version.

Fix bugs in proxy implementation when it comes to stream and pollset
handling and transfer lifetimes.

Curl_multi_xfer_sockbuf_borrow: work without multi

When a connection gets shutdown by a share, the easy handle
used is share->admin and it does not have a multi handle. In
that case let Curl_multi_xfer_sockbuf_borrow() allocate a
buffer to be freed on release.

This happens when a TLS filter sends its last notify through
a HTTP/3 proxy tunnel.
