+++ lib/http_proxy.h
@@ -68,8 +68,9 @@ CURLcode Curl_http_proxy_inspect_tunnel_response(
 
 CURLcode Curl_cf_http_proxy_insert_after(struct Curl_cfilter *cf_at,
                                          struct Curl_easy *data,
-                                         struct Curl_peer *dest,
-                                         uint8_t transport,
+                                         struct Curl_peer *peer,
+                                         struct Curl_peer *tunnel_peer,
+                                         uint8_t tunnel_transport,
                                          uint8_t proxytype);
 
 extern struct Curl_cftype Curl_cft_http_proxy;
