+++ lib/vquic/cf-ngtcp2-proxy.h
@@ -1,5 +1,5 @@
-#ifndef HEADER_CURL_H3_PROXY_H
-#define HEADER_CURL_H3_PROXY_H
+#ifndef HEADER_CURL_VQUIC_CF_NGTCP2_PROXY_H
+#define HEADER_CURL_VQUIC_CF_NGTCP2_PROXY_H
 /***************************************************************************
  *                                  _   _ ____  _
  *  Project                     ___| | | |  _ \| |
@@ -32,16 +32,19 @@
 
 CURLcode Curl_cf_ngtcp2_proxy_insert_after(struct Curl_cfilter *cf_at,
                                            struct Curl_easy *data,
-                                           struct Curl_peer *dest,
-                                           bool udp_tunnel);
+                                           struct Curl_peer *peer,
+                                           struct Curl_peer *tunnel_peer,
+                                           uint8_t tunnel_transport);
 
 CURLcode Curl_cf_ngtcp2_proxy_create(struct Curl_cfilter **pcf,
                                      struct Curl_easy *data,
+                                     struct Curl_peer *peer,
+                                     uint8_t transport_peer,
                                      struct connectdata *conn,
                                      struct Curl_sockaddr_ex *addr,
-                                     uint8_t transport_in,
-                                     uint8_t transport_out);
+                                     struct Curl_peer *tunnel_peer,
+                                     uint8_t tunnel_transport);
 
 #endif
 
-#endif /* HEADER_CURL_H3_PROXY_H */
+#endif /* HEADER_CURL_VQUIC_CF_NGTCP2_PROXY_H */
